requirements.txt
cryptography>=39.0.0
pynacl>=1.5.0 